api/requirements.txt
absl-py==1.3.0
alembic==1.8.1
aniso8601==9.0.1
astunparse==1.6.3
autopep8==2.0.0
beautifulsoup4==4.11.1
bidict==0.22.0
cachelib==0.9.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
deepface==0.0.75
filelock==3.8.0
fire==0.4.0
Flask==2.2.2
Flask-Cache==0.13.1
Flask-Caching==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==4.0.0
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.1
Flask-SQLAlchemy==3.0.2
flatbuffers==22.10.26
ftfy==6.1.1
gast==0.4.0
gdown==4.5.3
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
libclang==14.0.6
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
mtcnn==0.1.1
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
protobuf==3.19.6
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.9
simple-websocket==0.9.0
SimpleCache==0.1.2
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.1
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
